# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.ChatMessageListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/http.ChatMessageResponse'
        type: array
      pagination:
        $ref: '#/definitions/http.CursorPagination'
    type: object
  http.ChatMessageRequest:
    properties:
      client_message_id:
        type: string
      content:
        type: string
    type: object
  http.ChatMessageResponse:
    properties:
      client_message_id:
        type: string
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      edited_at:
        type: string
      id:
        type: string
      message_type:
        type: string
      sender_id:
        type: string
      sender_name:
        type: string
      sequence_number:
        type: integer
    type: object
  http.ChatMessageUpdateRequest:
    properties:
      content:
        type: string
    type: object
  http.ChatReadRequest:
    properties:
      sequence_number:
        type: integer
    type: object
...
  http.ConversationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/http.ConversationResponse'
        type: array
      pagination:
        $ref: '#/definitions/http.CursorPagination'
    type: object
  http.ConversationResponse:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      last_message_at:
        type: string
      last_message_preview:
        type: string
      last_message_sequence:
        type: integer
      status:
        type: string
      type:
        type: string
      unread_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  http.CursorPagination:
    properties:
      has_more:
        example: true
        type: boolean
      next_cursor:
        type: string
    type: object
...
info:
  contact: {}
  description: Public HTTP API exposed by the Book Store API Gateway.
  title: Book Store API
  version: "1.0"
paths:
...
  /api/v1/chat/conversations:
    get:
      parameters:
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Opaque cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ConversationListResponse'
      security:
      - BearerAuth: []
      summary: List accessible chat conversations
      tags:
      - Chat
  /api/v1/chat/conversations/{id}/messages:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Opaque cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ChatMessageListResponse'
      security:
      - BearerAuth: []
      summary: List conversation messages using cursor pagination
      tags:
      - Chat
    post:
      consumes:
      - application/json
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChatMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.ChatMessageResponse'
      security:
      - BearerAuth: []
      summary: Send an idempotent text message
      tags:
      - Chat
  /api/v1/chat/conversations/{id}/read:
    put:
      consumes:
      - application/json
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Last read sequence; zero means latest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChatReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Advance the current user's read cursor
      tags:
      - Chat
  /api/v1/chat/conversations/support:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.ConversationResponse'
      security:
      - BearerAuth: []
      summary: Open or return the current support conversation
      tags:
      - Chat
  /api/v1/chat/messages/{id}:
    delete:
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ChatMessageResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete an own message; admins may remove abusive messages
      tags:
      - Chat
    put:
      consumes:
      - application/json
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: New content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChatMessageUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ChatMessageResponse'
      security:
      - BearerAuth: []
      summary: Update an own chat message
      tags:
      - Chat
  /api/v1/chat/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Count unread chat messages
      tags:
      - Chat
  /api/v1/chat/ws:
    get:
      parameters:
      - description: One-time WebSocket ticket
        in: query
        name: ticket
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      summary: Upgrade to the chat WebSocket using a one-time ticket
      tags:
      - Chat
  /api/v1/chat/ws-ticket:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Issue a one-time short-lived WebSocket ticket
      tags:
      - Chat
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token using the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"